requirements.txt
PyCrypto
rethinkdb
redis
huey
bottle
scrypt
itsdangerous
